Conflicts:

	release/src-rt/linux/linux-2.6/include/net/netfilter/nf_conntrack.h
	release/src-rt/linux/linux-2.6/net/netfilter/xt_MARK.c
	release/src/router/rc/services.c

+++ release/src/router/rc/services.c
@@ -255,8 +255,10 @@ void start_dnsmasq()
 			}
 #endif
 		} else {
-			if (strcmp(nvram_safe_get(lanN_ifname),"")!=0)
+			if (strcmp(nvram_safe_get(lanN_ifname),"")!=0) {
+				fprintf(f, "interface=%s\n", nvram_safe_get(lanN_ifname));
 				fprintf(f, "no-dhcp-interface=%s\n", nvram_safe_get(lanN_ifname));
+			}
 		}
 	}
 	// write static lease entries & create hosts file
@@ -332,7 +334,11 @@ void start_dnsmasq()
 		}
 
 		if ((do_dhcpd_hosts > 0) && (*mac != 0) && (strcmp(mac, "00:00:00:00:00:00") != 0)) {
-			fprintf(f, "dhcp-host=%s,%s,%s\n", mac, ip, sdhcp_lease);
+			if (nvram_get_int("dhcpd_slt") == 0) {
+				fprintf(f, "dhcp-host=%s,%s\n", mac, ip);
+			} else {
+				fprintf(f, "dhcp-host=%s,%s,%s\n", mac, ip, sdhcp_lease);
+			}
 		}
 	}
 
@@ -798,7 +804,6 @@ void start_upnp(void)
 					char *lanip = nvram_safe_get(lanN_ipaddr);
 					char *lanmask = nvram_safe_get(lanN_netmask);
 					char *lanlisten = nvram_safe_get(upnp_lanN);
-
 					if((strcmp(lanlisten,"1")==0) && (strcmp(lanip,"")!=0) && (strcmp(lanip,"0.0.0.0")!=0)) {
 						fprintf(f,
 							"listening_ip=%s/%s\n",
@@ -913,7 +918,6 @@ void start_zebra(void)
 	char *wan_tx = nvram_safe_get("dr_wan_tx");
 	char *wan_rx = nvram_safe_get("dr_wan_rx");
 
-//	if ((*lan_tx == '0') && (*lan_rx == '0') && (*wan_tx == '0') && (*wan_rx == '0')) {
 	if ((*lan_tx == '0') && (*lan_rx == '0') && 
 		(*lan1_tx == '0') && (*lan1_rx == '0') && 
 		(*lan2_tx == '0') && (*lan2_rx == '0') && 
@@ -1035,6 +1039,7 @@ void start_syslog(void)
 	char s[64];
 	char cfg[256];
 	char *rot_siz = "50";
+	char *rot_keep = "1";
 	char *log_file_path;
 
 	argv[0] = "syslogd";
@@ -1052,17 +1057,23 @@ void start_syslog(void)
 	if (nvram_match("log_file", "1")) {
 		argv[argc++] = "-L";
 
+		if (strcmp(nvram_safe_get("log_file_size"), "") != 0) {
+			rot_siz = nvram_safe_get("log_file_size");
+		}
+		if (nvram_get_int("log_file_size") > 0) {
+			rot_keep = nvram_safe_get("log_file_keep");
+		}
+
 		// log to custom path - shibby
 		if (nvram_match("log_file_custom", "1")) {
 			log_file_path = nvram_safe_get("log_file_path");
-			argv[argc++] = "-s";
-			argv[argc++] = "5000";
-			argv[argc++] = "-b";
-			argv[argc++] = "5";
+			argv[argc++] = rot_siz;
 			argv[argc++] = "-O";
 			argv[argc++] = log_file_path;
-			remove("/var/log/messages");
-			symlink(log_file_path, "/var/log/messages");
+			if (strcmp(nvram_safe_get("log_file_path"), "/var/log/messages") != 0) {
+				remove("/var/log/messages");
+				symlink(log_file_path, "/var/log/messages");
+			}
 		}
 		else
 
@@ -1102,6 +1113,10 @@ void start_syslog(void)
 		if (isdigit(*b_opt)) {
 			argv[argc++] = "-b";
 			argv[argc++] = b_opt;
+		} else
+		if (nvram_get_int("log_file_size") > 0) {
+			argv[argc++] = "-b";
+			argv[argc++] = rot_keep;
 		}
 	}
 
@@ -1249,43 +1264,83 @@ void stop_ntpc(void)
 
 static void stop_rstats(void)
 {
-	int n;
+	int n, m;
 	int pid;
+	int pidz;
+	int ppidz;
+	int w = 0;
 
 	n = 60;
+	m = 15;
 	while ((n-- > 0) && ((pid = pidof("rstats")) > 0)) {
-		if (kill(pid, SIGTERM) != 0) break;
+		w = 1;
+		pidz = pidof("gzip");
+		if (pidz < 1) pidz = pidof("cp");
+		ppidz = ppid(ppid(pidz));
+		if ((m > 0) && (pidz > 0) && (pid == ppidz)) {
+			syslog(LOG_DEBUG, "rstats(PID %d) shutting down, waiting for helper process to complete(PID %d, PPID %d).\n", pid, pidz, ppidz);
+			--m;
+		} else {
+			kill(pid, SIGTERM);
+		}
 		sleep(1);
 	}
+	if ((w == 1) && (n > 0))
+		syslog(LOG_DEBUG, "rstats stopped.\n");
 }
 
 static void start_rstats(int new)
 {
 	if (nvram_match("rstats_enable", "1")) {
 		stop_rstats();
-		if (new) xstart("rstats", "--new");
-			else xstart("rstats");
+		if (new) {
+			syslog(LOG_DEBUG, "starting rstats (new datafile).\n");
+			xstart("rstats", "--new");
+		} else {
+			syslog(LOG_DEBUG, "starting rstats.\n");
+			xstart("rstats");
+		}
 	}
 }
 
 static void stop_cstats(void)
 {
-	int n;
+	int n, m;
 	int pid;
+	int pidz;
+	int ppidz;
+	int w = 0;
 
 	n = 60;
+	m = 15;
 	while ((n-- > 0) && ((pid = pidof("cstats")) > 0)) {
-		if (kill(pid, SIGTERM) != 0) break;
+		w = 1;
+		pidz = pidof("gzip");
+		if (pidz < 1) pidz = pidof("cp");
+		ppidz = ppid(ppid(pidz));
+		if ((m > 0) && (pidz > 0) && (pid == ppidz)) {
+			syslog(LOG_DEBUG, "cstats(PID %d) shutting down, waiting for helper process to complete(PID %d, PPID %d).\n", pid, pidz, ppidz);
+			--m;
+		} else {
+			kill(pid, SIGTERM);
+		}
 		sleep(1);
 	}
+	if ((w == 1) && (n > 0))
+		syslog(LOG_DEBUG, "cstats stopped.\n");
 }
 
 static void start_cstats(int new)
 {
 	if (nvram_match("cstats_enable", "1")) {
 		stop_cstats();
-		if (new) xstart("cstats", "--new");
-			else xstart("cstats");
+		if (new) {
+			syslog(LOG_DEBUG, "starting cstats (new datafile).\n");
+			xstart("cstats", "--new");
+		} else {
+			syslog(LOG_DEBUG, "starting cstats.\n");
+			xstart("cstats");
+		}
 	}
 }
 
@@ -1455,7 +1510,8 @@ static void start_ftpd(void)
 	if ((buf = strdup(nvram_safe_get("ftp_users"))) != NULL)
 	{
 		/*
-		username<password<rights[<root_dir]
+		username<password<rights[<root_dir>]
+
 		rights:
 			Read/Write
 			Read Only
@@ -1872,6 +1928,10 @@ static void start_nas_services(void)
 #ifdef TCONFIG_MEDIA_SERVER
 	start_media_server();
 #endif
+#ifdef TCONFIG_UPS
+	start_ups();
+#endif
+
 }
 
 static void stop_nas_services(void)
@@ -1890,6 +1950,10 @@ static void stop_nas_services(void)
 #ifdef TCONFIG_SAMBASRV
 	stop_samba();
 #endif
+#ifdef TCONFIG_UPS
+	stop_ups();
+#endif
+
 }
 
 void restart_nas_services(int stop, int start)
@@ -2121,9 +2185,9 @@ TOP:
 		goto CLEAR;
 	}
 
-	if (strcmp(service, "account") == 0) {
-		if (action & A_STOP) stop_account();
-		if (action & A_START) start_account();
+	if (strcmp(service, "arpbind") == 0) {
+		if (action & A_STOP) stop_arpbind();
+		if (action & A_START) start_arpbind();
 		goto CLEAR;
 	}
 
@@ -2147,12 +2211,6 @@ TOP:
 	if (action & A_START) {
 		new_qoslimit_start();
 	}
-		goto CLEAR;
-	}
-
-	if (strcmp(service, "arpbind") == 0) {
-		if (action & A_STOP) stop_arpbind();
-		if (action & A_START) start_arpbind();
 		goto CLEAR;
 	}
 
@@ -2494,6 +2552,14 @@ TOP:
 	}
 #endif
 
+#ifdef TCONFIG_UPS
+	if (strcmp(service, "ups") == 0) {
+		if (action & A_STOP) stop_ups();
+		if (action & A_START) start_ups();
+		goto CLEAR;
+	}
+#endif
+
 #ifdef TCONFIG_USB
 	// !!TB - USB Support
 	if (strcmp(service, "usb") == 0) {
@@ -2598,7 +2664,13 @@ static void do_service(const char *name, const char *action, int user)
 
 	snprintf(s, sizeof(s), "%s-%s%s", name, action, (user ? "-c" : ""));
 	nvram_set("action_service", s);
-	kill(1, SIGUSR1);
+
+	if (nvram_match("debug_rc_svc", "1")) {
+		nvram_unset("debug_rc_svc");
+		exec_service();
+	} else {
+		kill(1, SIGUSR1);
+	}
 
 	n = 150;
 	while (nvram_match("action_service", s)) {
